+++ includes/main_menu.h
@@ -1,0 +1,11 @@
+#pragma once
+
+// --- --- main menu functions --- ---
+void main_menu();
+void main_menu_text();
+void register_new_client();
+void remove_client();
+void list_active_clients();
+void client_management_menu();
+void sort_by_total_spent();
+void save_as_csv_menu();
